Moved coursier specific code into separate trait/object
